Added bot files

+++ api_store/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, func
+from sqlalchemy import Column, Integer, String, DateTime, func, BigInteger
 from sqlalchemy.orm import relationship
 
 from ..db.database import Base
@@ -10,7 +10,7 @@ class User(Base):
     user_id = Column(Integer, primary_key=True, autoincrement=True)
     full_name = Column(String(255), nullable=True)
     address = Column(String(255), nullable=True)
-    tg_id = Column(Integer, nullable=False)
+    tg_id = Column(BigInteger, nullable=False)
     created_at = Column(DateTime, default=func.now())
 
     orders = relationship("Order", back_populates="user")
